move index store types
